refactor(openclaw): split onboarding conversation content from mechanics
agent-onboarding-config.ts now holds everything the bot says or offers
during group setup — the purpose prompt, footer, and cards; the topic
prompt, submit label, and per-purpose suggestions. agent-onboarding.ts
keeps the operations: turning that content into A2UI blobs, the offer
gates, and the group/channel scries. Copy and option changes no longer
touch operational code.
